Add new view

+++ CurrencyConverter/Models/Chart/Ticks.cs
@@ -5,6 +5,6 @@
         public bool beginAtZero { get; set; }
         public bool display { get; set; }
 
-        public int stepSize { get; set; }
+        public double stepSize { get; set; }
     }
 }
